[to fix] Remove evaluation of blog preview
By evaluating the blog preview a situation may arise where an html tag
is open, evaluated, but never close and thus raise an error. I need to
fix this somehow since I don't want raw html in the preview, but that
is better than an error for the time being.