[app] Functional add_post route
In this commit I convert the add_post.py script into a function to be
found in utils.py, that is used by the add_post method to generate a
post from the title, date and markdown provided in the html
form. Still missing:

1- Security: some sort of password so not everyone can add posts
2- Images: possibility to upload images and store them in the correct
location assumed by utils.add_posts.
3- Nice text input form: make the form for the markdown a little bit
better.