[app] Reorganize
Pretty crazy organization taken from this guy:
https://blog.miguelgrinberg.com/post/the-flask-mega-tutorial-part-iii-web-forms.

At the end of the day the "app.py" that flask requires needs to have a
variable app with is an instance of flask.Flask and has been
configured appropriately (i.e. routes have been added, and these
routes and their dependencies work). So what we do is define a package
called app with an __init__.py which defines the app variable and
importers the routes which, together with all their dependencies, are
in this app package. Furthermore, at the root of the directory we
define a .py file that imports the app variable from the app package,
and thus satisfies the requirements for an "app.py" to be used by
flask. All veeery nicely separated.