[env] Replace dotenv with config.py
Since 8bc54e8, we make extensive use of config.py, which is a standard
flask configuration file. Since it is can define default values for
config variables the example.env and subsequent .env logic seems
unnecessary (i.e. the role of example.env is fulfilled by the defaults
and the .env needs to be written from scratch anyway).