[db] Initialize database on app start up
If the table already exists in a posts.db file in the correct
location, it is used as the posts table. The only issue with this is
if the table does not satisfy the necessary schema, in which case an
internal server error will show up when a post gets added. It would be
nice to check this in the future.